Rename: HttpMethod(enum) 디렉토리 이동
