+++ src/main/java/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package util.reader;
+package request;
 
 import java.util.Arrays;
